Refactor shift and algebraic simplifications/transforms
Removed nested shift folding and add/sub algebraic simplifications. Added new shift simplification rules and transforms for out-of-range shift amounts (ShiftLeft/Right32/64). Updated transform registration and project structure. Cleaned up related JSON and C# files.
