+++ Source/Mosa.Compiler.Framework/Mosa.Compiler.Framework.csproj
@@ -5,4 +5,8 @@
 	  <ProjectReference Include="..\Mosa.Utility.Configuration\Mosa.Utility.Configuration.csproj" />
   </ItemGroup>
 
+  <ItemGroup>
+    <Folder Include="Transforms\Optimizations\Auto\" />
+  </ItemGroup>
+
 </Project>
