+++ Source/Mosa.Compiler.Framework/Transforms/Optimizations/Auto/AutoTransforms.cs
@@ -245,14 +245,6 @@ public static class AutoTransforms
 		new Simplification.Compare32x64SameAndNotEqual(),
 		new Simplification.Compare64x32SameAndNotEqual(),
 		new Simplification.Compare64x64SameAndNotEqual(),
-		new Simplification.Add32v1(),
-		new Simplification.Add32v1_v1(),
-		new Simplification.Add64v1(),
-		new Simplification.Add64v1_v1(),
-		new Simplification.Sub32v1(),
-		new Simplification.Sub32v1_v1(),
-		new Simplification.Sub64v1(),
-		new Simplification.Sub64v1_v1(),
 		new Simplification.Compare32x32SwapToZero(),
 		new Simplification.Compare32x32SwapToZero_v1(),
 		new Simplification.Compare32x32SwapToZero_v2(),
@@ -356,6 +348,10 @@ public static class AutoTransforms
 		new Simplification.MulCarryOut64CarryNotUsed(),
 		new Simplification.MulOverflowOut32CarryNotUsed(),
 		new Simplification.MulOverflowOut64CarryNotUsed(),
+		new Simplification.ShiftRight32(),
+		new Simplification.ShiftRight64(),
+		new Simplification.ShiftLeft32(),
+		new Simplification.ShiftLeft64(),
 		new StrengthReduction.Add32Zero(),
 		new StrengthReduction.Add32Zero_v1(),
 		new StrengthReduction.Add64Zero(),
@@ -720,10 +716,6 @@ public static class AutoTransforms
 		new ConstantFolding.SubAddR4_v1(),
 		new ConstantFolding.SubAddR8(),
 		new ConstantFolding.SubAddR8_v1(),
-		new ConstantFolding.ShiftLeft32x2(),
-		new ConstantFolding.ShiftLeft64x2(),
-		new ConstantFolding.ShiftRight32x2(),
-		new ConstantFolding.ShiftRight64x2(),
 		new ConstantFolding.Load32FoldAdd32(),
 		new ConstantFolding.Load32FoldAdd64(),
 		new ConstantFolding.Load64FoldAdd32(),
