--- Source/Mosa.Compiler.Framework/Transforms/Optimizations/Auto/ConstantFolding/ShiftLeft64x2.cs
@@ -1,51 +1,0 @@
-// Copyright (c) MOSA Project. Licensed under the New BSD License.
-
-// This code was generated by an automated template.
-
-namespace Mosa.Compiler.Framework.Transforms.Optimizations.Auto.ConstantFolding;
-
-public sealed class ShiftLeft64x2 : BaseTransform
-{
-	public ShiftLeft64x2() : base(IR.ShiftLeft64, TransformType.Auto | TransformType.Optimization, 90)
-	{
-	}
-
-	public override bool Match(Context context, Transform transform)
-	{
-		if (!context.Operand1.IsVirtualRegister)
-			return false;
-
-		if (!context.Operand1.IsDefinedOnce)
-			return false;
-
-		if (context.Operand1.Definitions[0].Instruction != IR.ShiftLeft64)
-			return false;
-
-		if (!IsResolvedConstant(context.Operand1.Definitions[0].Operand2))
-			return false;
-
-		if (!IsResolvedConstant(context.Operand2))
-			return false;
-
-		if (!IsIntegerBetween0And64(context.Operand1.Definitions[0].Operand2))
-			return false;
-
-		if (!IsIntegerBetween0And64(context.Operand2))
-			return false;
-
-		return true;
-	}
-
-	public override void Transform(Context context, Transform transform)
-	{
-		var result = context.Result;
-
-		var t1 = context.Operand1.Definitions[0].Operand1;
-		var t2 = context.Operand1.Definitions[0].Operand2;
-		var t3 = context.Operand2;
-
-		var e1 = Operand.CreateConstant(Add64(To64(t2), To64(t3)));
-
-		context.SetInstruction(IR.ShiftLeft64, result, t1, e1);
-	}
-}
