--- Source/Mosa.Compiler.Framework/Transforms/Optimizations/Auto/Simplification/Add32v1.cs
@@ -1,71 +1,0 @@
-// Copyright (c) MOSA Project. Licensed under the New BSD License.
-
-// This code was generated by an automated template.
-
-namespace Mosa.Compiler.Framework.Transforms.Optimizations.Auto.Simplification;
-
-public sealed class Add32v1 : BaseTransform
-{
-	public Add32v1() : base(IR.Add32, TransformType.Auto | TransformType.Optimization)
-	{
-	}
-
-	public override bool Match(Context context, Transform transform)
-	{
-		if (!context.Operand1.IsVirtualRegister)
-			return false;
-
-		if (!context.Operand1.IsDefinedOnce)
-			return false;
-
-		if (context.Operand1.Definitions[0].Instruction != IR.Sub32)
-			return false;
-
-		if (!AreSame(context.Operand1.Definitions[0].Operand2, context.Operand2))
-			return false;
-
-		return true;
-	}
-
-	public override void Transform(Context context, Transform transform)
-	{
-		var result = context.Result;
-
-		var t1 = context.Operand1.Definitions[0].Operand2;
-
-		context.SetInstruction(IR.Move32, result, t1);
-	}
-}
-
-public sealed class Add32v1_v1 : BaseTransform
-{
-	public Add32v1_v1() : base(IR.Add32, TransformType.Auto | TransformType.Optimization)
-	{
-	}
-
-	public override bool Match(Context context, Transform transform)
-	{
-		if (!context.Operand2.IsVirtualRegister)
-			return false;
-
-		if (!context.Operand2.IsDefinedOnce)
-			return false;
-
-		if (context.Operand2.Definitions[0].Instruction != IR.Sub32)
-			return false;
-
-		if (!AreSame(context.Operand1, context.Operand2.Definitions[0].Operand2))
-			return false;
-
-		return true;
-	}
-
-	public override void Transform(Context context, Transform transform)
-	{
-		var result = context.Result;
-
-		var t1 = context.Operand1;
-
-		context.SetInstruction(IR.Move32, result, t1);
-	}
-}
