+++ Source/Mosa.Compiler.Framework/Transforms/Optimizations/Auto/Simplification/ShiftLeft32.cs
@@ -1,0 +1,38 @@
+// Copyright (c) MOSA Project. Licensed under the New BSD License.
+
+// This code was generated by an automated template.
+
+namespace Mosa.Compiler.Framework.Transforms.Optimizations.Auto.Simplification;
+
+public sealed class ShiftLeft32 : BaseTransform
+{
+	public ShiftLeft32() : base(IR.ShiftLeft32, TransformType.Auto | TransformType.Optimization, 25)
+	{
+	}
+
+	public override bool Match(Context context, Transform transform)
+	{
+		if (IsResolvedConstant(context.Operand1))
+			return false;
+
+		if (!IsResolvedConstant(context.Operand2))
+			return false;
+
+		if (!IsGreaterOrEqual(To32(context.Operand2), 32))
+			return false;
+
+		return true;
+	}
+
+	public override void Transform(Context context, Transform transform)
+	{
+		var result = context.Result;
+
+		var t1 = context.Operand1;
+		var t2 = context.Operand2;
+
+		var e1 = Operand.CreateConstant(Sub32(And32(To32(t2), 32), 1));
+
+		context.SetInstruction(IR.ShiftLeft32, result, t1, e1);
+	}
+}
