+++ Source/Mosa.Compiler.Framework/Transforms/Optimizations/Auto/Simplification/ShiftRight64.cs
@@ -1,0 +1,38 @@
+// Copyright (c) MOSA Project. Licensed under the New BSD License.
+
+// This code was generated by an automated template.
+
+namespace Mosa.Compiler.Framework.Transforms.Optimizations.Auto.Simplification;
+
+public sealed class ShiftRight64 : BaseTransform
+{
+	public ShiftRight64() : base(IR.ShiftRight64, TransformType.Auto | TransformType.Optimization, 25)
+	{
+	}
+
+	public override bool Match(Context context, Transform transform)
+	{
+		if (IsResolvedConstant(context.Operand1))
+			return false;
+
+		if (!IsResolvedConstant(context.Operand2))
+			return false;
+
+		if (!IsGreaterOrEqual(To64(context.Operand2), 64))
+			return false;
+
+		return true;
+	}
+
+	public override void Transform(Context context, Transform transform)
+	{
+		var result = context.Result;
+
+		var t1 = context.Operand1;
+		var t2 = context.Operand2;
+
+		var e1 = Operand.CreateConstant(Sub64(And64(To64(t2), 64), 1));
+
+		context.SetInstruction(IR.ShiftRight64, result, t1, e1);
+	}
+}
